requirements.txt
Jinja2==2.10.3
redis==3.3.11
google-cloud-storage==1.23.0
google-cloud-kms==1.2.1
